move pjson out of shared dir
